Move NewTabLink to "common/components/ui" instead of having its own folder
