+++ common/components/ui/index.js
@@ -5,3 +5,4 @@ export { default as Identicon } from './Identicon';
 export { default as Modal } from './Modal';
 export { default as UnlockHeader } from './UnlockHeader';
 export { default as QRCode } from './QRCode';
+export { default as NewTabLink } from './NewTabLink.jsx';
